forge/requirements.txt
transformers
torch
sentence-transformers
pyyaml
aiohttp
beautifulsoup4
ratelimit
numpy
sqlalchemy
